feat: add initial protobuf files and converters
to begin with, adding the basic message formats we plan to pass
around between nodes in our framework. the data will be serialized
using protobuf, so here we add the main components.

additionally, we have a module array_converters.py for converting
from our protobuf 'Scan2d' to xarray or sidpy (formats we expect
the scan to be read in). this is unit tested.