[feat] add pub-sub and cache logic
here, we are adding the overall logic for a publisher, subscriber,
and an intermediate cache in between them.

In our afspm system, we expect:
- a DeviceController to have a publisher class, which it uses to send
state changes and scans out to any subscribers connected.
- an AfspmController, which subscribes to the DeviceController and
implements a PubSubCache. This PubSubCache acts as an 'intermediary',
to cache data received. It's purpose is to be able to get any subscriber
up-to-date if they join late (or crash and are restarted). Without it,
we could run into a scenario where an Analyzer node crashes and, on
restart, does not have enough information to analyze.
- 0 or more subscriber nodes connected to our PubSubCache's publisher
end. These nodes could do any number of things: visualizing data as we
run scans, analyzing data to try and detect issues (e.g. tip condition
deterioration), or even simply detecting that the scan we requested has
finished.

These 3 classes (publisher, subscriber, pubsubcache), encapsulate this
logic. They rely heavily on zmq for communication between them, and
protobuf to serialize/deserialize our defined classes.

Lastly, we have an added module cache_logic, which contains additional
helpers for creating caches based on desired criteria.

All the above is tested in the 2 new test modules.