[refactor] modify package directory structure
we have done the following:
- protos are now in ./io/protos/src (instead of ./interfaces), as they
are logically part of the i/o.
- compiled protos are now in ./io/protos/generated (instead of
./generated/python), for the same reason. Note that we do not explicit
python anymore; I'm not sure it's necessary, and it just adds an extra
folder.
- publisher.py, subscriber.py, and pubsubcache.py are now in
./io/pubsub.
- cache_logic.py has been split: the abstract CacheLogic class (and the
non-class accessors) stay; ProtoBasedLogic and PBCWithROILogic are moved
to pbc_logic.py. Both of these are placed in ./io/cache.

This was all done in an attempt to properly separate all these classes.
We will be working on the 'control' section shortly, which will entail
another set of classes in ./io/control.