feat: add DeviceController
An initial draft of the DeviceController is added. It is an abstract
class that handles the main polling logic (a) with a client (via the
ControlServer instance, which communicates with a ControlRouter); and (b)
with the device itself, to determine its current state. When changes are
detected, we send them out via the Publisher instance.

Testing and missing components will be in a subsequent CL.