feat: add heartbeat logic
This CL adds heartbeat logic, which should allow us to detect afspm
components crashing/freezing (and presumably restart them). It involves
two simple classes: Heartbeater and HeartbeatListener.

The assumption is that any AfspmComponent (or node?) will contain a
Heartbeater, and will be spawned linked to a HeartbeatListener
thread/process. If the listener detects that the component has
frozen/crashed, it will respawn it.