fix/refactor: minor changes in heartbeat.py
The main fixes here:
- reset_received_kill_signal() was changed to reset(), because we
realized that other internal logic needs to be reset when we respawn the
associated Heartbeat process.
- check_if_dead's default timeout_ms was pulled *way* down, to just 10
ms. It was interfering with other checks. Since all of our afspm
components have a separate sleep, I'm pretty OK with making it fairly
fast. In principle, we could make it incredibly fast (1 ms!?)