feat: add afspm_component logic
In this CL, we add the two backbone afspm component classes:
- AfspmComponent: parent class, which contains heartbeat logic, so we
can know when a component is frozen/crashes.
- AfspmComponentsMonitor: a helper class that functions as a factory and
monitor of a list of components we wish to spawn on a PC. It will spawn
them and, for each, create a linked HeartbeatListener to monitor for
life. If a component crashes/freezes, this class will restart them. If a
component purposefully exists, it will close the associated listener.

We have a unit test that validates this logic.