refactor: modify Scan2d attributes and names
In a (potentially misguided) attempt to clean up the ScanParameters2d
message, we pulled the 'data' attributes out from there and put them.
along with the raw data, into a ScanData message.

So now, Scan2d consists just of the
physical *params* (ScanParameters2d), and the *data* (ScanData).

We also added an optional timestamp attribute in ScanParameters2d. The
idea is for different DeviceController's to set this attribute based on
the file creation time (we will provide a helper), and to fill that
parameter. This way, determination of whether a new scan was created is
somewhat simplified.