feat: add KILL signal logic to pubsub
Here, we add KILL signal logic to the pubsub constituents, so we can
tell different components when an experiment ends. Particularly:
- PubSubCache can now send a message provided by its
owner/user (presumably, the kill signal; though we leave it open).
- Subscriber explicitly checks for a defined KILL_SIGNAL whenever
polling. If so, it sets an internal variable to indicate a shutdown was
requested.

We have updated the tests to test this, as well as clean up the code a
bit (using more fixtures to minimize code duplication).