feat: add 'admin' capabilities to clients
Here, we add two key capabilities to clients:
- The ability to call for the end of an experiment;
- The ability to set the current ControlMode.

Note that these controls are considered 'admin' controls, in that only
special AfspmComponents should use them! To simplify this, we have
created a child class of ControlClient called AfspmControlClient. This
is the class that has methods for these commands. However, note that the
actual ControlRequests are not hidden via inheritance; they are simply
added to ControlRequest in control.proto.

Additionally, we 'refactor' ControlState slightly:
- problems_list is now problems_set, as it is elsewhere in the code.
- under_control shifts to client_in_control_id, now a string instead of
a boolean. If the string is empty, no client is assumed to be under
control.

Note that in control_router, we:
- Added control logic to handle the new requests;
- Added a get_control_state() method, to package the current control
state into a ControlState message.
- Fixed the typing for 'client' in many methods: it was int, but has
been a str for a bit now (see _parse_client_id()).