feat: add expected __init__.py files for package
This package is (presumably) missing __init__.py files in all
directories, for it to be properly handled as one. It appears that
the reason this has not caused issues earlier is due to Python 3.3's
introduction of 'namespace' packages. Namespace packages do not require
the __init__.py; they are used explicitly for the case where you want to
set global, inter-package module hierarchies. For example, if you have
two separate packages, A and B, which both use namespace 'google.cloud',
your structure would look like this:

A/
    google/  <-- namespace package
        cloud/  <-- namespace package
            A/  <-- standard package
                __init__.py

B/
    google/  <-- namespace package
        cloud/  <-- namespace package
            B/  <-- standard package
                __init__.py

Thus, they would both exist within the google.cloud 'namespace'. Nifty.

But: this does not apply to us! All our logic is inside this repo. So we
will stick with standard packages and have __init__.py's everywhere.