refactor: update protos and update elsewhere
In a previous CL, I decoupled the data 'parameters' (shape and units)
out of ScanParameters2d, and into a ScanData struct, with the actual
data. While well intentioned, this was *wrong*: we need to keep the data
parameters in ScanParameters2d, because that is how the user changes the
... scan parameters! By decoupling it, the user had no way to change the
resolution of their scans.

Thus, we move stuff around yet again:
- data shape and units is in a DataAspects struct;
- spatial roi and units are in a SpatialAspects struct;
- both of the above are in ScanParameters2d;
- The data values are returned to Scan2d.
- the timestamp parameter is moved to Scan2d as well, since it's not
something the user would set!

(pretty much) everything else in this CL is just changing tests/rest of
the code to account for these protos changes.

Note: in this CL we *also* create a ScanStateMsg message, to hold the
ScanState. This is because, unlike the control data, we cannot send
enums alone via the SUB/PUB design we have set up (due to how the cache
is architected). Thus, we will need to wrap it in this message to pass
it around.