feat: add components logic and tests
In this CL, we introduce the 3 main component-level classes:
- AfspmComponent: the base class for any component in our system.
- AfspmController: The intermediary between all components and the
DeviceController, which talks directly to the SPM system.
- DeviceController: the individual class that talks to the SPM system.

For AfspmComponent: we changed it to no longer be an abstract class!
Now, there are 2 general methods you can choose to override for special
logic:
- run_per_loop(), if you want to add any general logic to perform on an
iteration through our larger run() loop.
- on_message_received(), if you want to do anything on a message being
received from the publisher.

test_components_logic.py tests the larger logic of the system, with a
sample DeviceController from sample_components.py.