fix: tmp hack to get pytest running
Pytest does not play super well with zmq on shutdown. In fact, it does
not explicitly close any sockets between tests! I've mostly worked
around this by forcing 'bind' sockets to close in threads, but it does
not seem to work 100% of the time.

Here, I've added a hack-ey delay after ending the experiment, to try to
minimize errors. I should revisit this later, as it is not great.