fix: update pubsub logic tests to pass more often
Isn't that a scary statement? pytest and zmq do not seem to behave
consistently with threads, because pytest does not seem to properly
destroy objects upon a thread's closure. Thus, although we try our best
to be explicit in closing appropriate sockets between tests, not
everything is hunky-dory.

In this CL, we try to clean up test_pubsub_logic to pass more
consistently. There is still one funky sleep we *need* to put in
place (on startup of the pubsubcache thread), for the tests to pass
consistently. As of today, the *why* is sadly a mystery (why is this not
necessary for the threads in test_components_logic.py?), but I will
stick with it for now.

Additionally, we have updated pubsubcache's poll() method. Now, it
handles subscriptions *before* handling new messages coming in. The
reason: if we sent messages first, we would
- Add the new message to our cache and send it out;
- Check new subscriptions; if we see one, send the last thiing in our
cache!

However, if you had a new subscription on *the same loop* where you got
a new message, you ran the risk of sending the same message 2x: first,
right after adding it to the cache; and second, when you realized you
had a new subscription. By changing the order, this is avoided.