feat: add kwargs to all classes, for instantiation
Our top-level usage of afspm will involve instantiating components from
a TOML dictionary. We will have a bunch of parsing methods to handle
this easily for the user, but at the end we will end up with a big dict
with many parameters. For a given class we wish to instantiate, it is
not a guarantee that the kwargs dict we provide will *only* have the
input arguments defined. As a simple example, a user could have provided
input parameters that did not exist for a class in this TOML config.

To avoid having a method that explicitly parses out the necessary
arguments for a class's constructor from a kwargs dict, we simply add
kwargs as an input to all our constructors.

Note: we have not done this for AfspmComponentsMonitor in this CL,
because we are in the process of updating that class. The addition of
kwargs will be in that separate CL for that class.