feat: add dict parsing/populating logic
The final 'top-level' running of afspm is expected to be based off of a
TOML config file. These will be read in as a dict containing multiple
levels of key:val pairs; many of these will be dicts themselves! So
effectively, we have a multi-level dict we need to parse.

There are 3 levels to this:
1. We want to allow variables to be provided by the user in their config
file, and reused. For example:
      pub_url = 'tcp://127.0.0.1:5555'
      [...]
      [MyTestPublisher]
      url = 'pub_url'
We should be able to 'expand' these variables, so that we end up with a
dict where these references are replaced by the actual objects. This is
handled by parser.expand_variables_in_dict().
2. We need to be able to reference classes for import or instantiation
within the config. For example:
      sub_extract_proto =
      'afspm.io.cache.cache_logic.CacheLogic.extract_proto'
We therefore need to be able to 'evaluate' these strings, to import them
and (potentially, if they end in ()) instantiate them. This is handled
by _evaluate_values_recursively(), which uses _evaluate_value_str(), which uses _import_from_string() (in parser).
3. We want to be able to instantiate classes with multiple arguments
from the config (as opposed to in (2), where we have simple no-arg
constructors via '()'). For example:
    class: 'afspm.components.afspm_component.AfspmComponent',
    loop_sleep_s: 0,
    hb_period_s: 5,
    [...]
We should be able to take a dict like this, containig
- a 'class' key indicating the class to instantiate; and
- the other key:val pairs as the class constructor arguments;
and instantiate it. This is handled within _instantiate_classes_recursively().

Note that (2) and (3) are bundled into a method
construct_and_run_component(), because we want to allow importing and
instantiation of all of these objects from *within* a new
thread/process. This is what AfspmComponentsMonitor uses to spawn each component.

All of the above logic is defined in afspm/utils/parser.py. We have the
main logic tested in tests/utils/test_parser.py.

Lastly, we modified AfspmComponentsMonitor significantly to account for
usage of this new parsing logic. Previously, it expected to receive a
list of (constructor, kwargs_dict) tuples; however, to ensure all
objects were instantiated *within* their process/thread, we modified the
logic to just take in a dictionary containing component_name:params_dict
key:val pairs. We also modified
tests/components/test_afspm_components_monitor.py accordingly.

Note also that we added the kwargs input argument to
AfspmComponentsMonitor, so we can spawn it from a dict with extra
args (as in the prior CL).