fix: hack fix to minimize UT fail
Like other zmq-based UTs, it seems stuff is not closed properly on test
closure. Thus, we have to explicit closing stuff and add ugly timed
sleeps when we tell the experiment to end.

One of these tests' sleeps is not *quite* long enough, so it crahes
sometimes. This CL just increases its duration.

Long-term, we need to fix these UTs so they don't have hack-ey sleeps.