feat: add top-level spawn module
This CL adds our top-level accessor, spawn.py. In it, we have our main
method spawn_components(), which takes in a TOML config file and a list
of components to spawn from it.

We expect users to almost exclusively use this accessor when running
experiments. It is here that you can start up a components monitor (and
associated components) on a particular device, based on the provided
config.

We also add test_spawn, to test some of the internal methods that are
not explicitly tested elsewhere.