fix: ensure heartbeat urls are in temp dir
It appears there is a weird bug in zmq (which according to their github
has already been resolved as of Feb 18 2019, tied to issue #3387): ipc
sockets do not delete their file descriptors on exit. This means that
any of the socket 'urls' we use that are ipc will create a 0-byte file
in the directory we pass via url:
   'ipc://path/to/file/filename'
Since we are using this for our heartbeat urls, we need to have a
cleaner workaround. For now, we are using python's tempfile package to
place all of these file descriptors in the main 'tmp' directory in a
cross-platform way.