feat/fix: Improve parser features
While testing the parser, I realized certain functionality was missing
and there were some bugs. Particularly:
- _expand_variables_recursively() was not expanding all variables. The
issue was that we were recursing, checking all variables for the
top-level equivalents in config_dict. However, we had not replaced
nested dicts in config_dict, so we were copying unexpanded dicts over
in certain scenarios!. The easy fix was to do a deep copy of the
config_dict at the beginning, and edit it directly as we walked through
it.
- in _evaluate_str(), we upgraded it to parse method arguments and
expand them recursively. Thus, something like method(a, method2(), c)
would get properly expanded out. Before this, we were *not* doing this
and it was breaking with cache logic methods.