feat/fix: fix bugs device/afspm controllers
AfspmController and DeviceController had hardcoded names. This made some
sense, but actually broke our heartbeat logic! So, to avoid this, we
remove this logic and allow the user to give them whatever name they
want (as set in the TOML key).

Also, in afspm_controller::_handle_shutdown(), we explicitly call
heartbeater.handle_closing(). Otherwise, the components monitor does not
know that we meant to close, and restarts!