feat: refactor spawn
Main things we have done:
- Remove _get_monito_parameters(), in favor of just looking for the
monitor key and passing the whole params dict to the constructor (as
kwargs).
- Removed _prepare_dict_for_spawning(), as we now use the two internal
methods within the main spawn_components() call.
- We copy the TOML 'name' key into its sub dict before storing that
component dict (for spawning). This allows us to spawn stuff
properly (with their names!).
- In logging, we currently hardcoded the 'root' logger as 'afspm',
rather than expliciting the *actual* root logger. I'm not sure if this
is what we will do in the end: it has the advantage of removing 3rd
party logs (which can be quite a lot); on the other hand, it means that
any non-afspm loggers we use will *not* be logged!