fix: update control_router for out-of-time reply
Similar to control_client, we need a way to back out if our request does
not get a reply in time. Without this, we could break ZMQ by not waiting
for a reply long enough, and then sending a new request before we have
received a reply.

To do this, we add _init_backend() and _close_backend() methods.
Additionally, we modify the logging for this kind of error to
logger.error, to indicate its severity.