fix: change some list typing to be tuples
There were a couple of cases of typing where I put something like
list[float, int]. This is not allowed and will likely throw an error
whenever we actually do any typing checks. For now, I've switched them
to tuple[float, int]. This may cause issues, as we are likely dealing
with lists in most cases (due to spawning from a TOML). We will deal
with this when it comes up (I primarily view typing as providing hints
for developers as to *what* should go in),