feat: create_roi_proto_hist_list takes 2D size
Previously, it only took a 1D size, assuming the size was equivalent
across dimensions.

Doing so highlighted that the toml package does not support complicated
list hierarchies, although TOML does. The tomli package does (this is
what got integrated into Python 3.11), so we switch to it.

tomli reads files as binary (does not take an encoding parameter), so we
remove that from spawn.py as well.