fix/feat: add control_mode_to_run to scan_handler
We want to stop sending control requests if we are not in control
and *not* in the 'desired' ControlMode. Thus, we add a
control_mode_to_run parameter; if the controller is not in this mode, we
effectively do nothing.

In _perform_scanning_logic(), we handle the early return of the
ControlMode not being correct in conjunction with the early return for
when the ScanStates are undefined. In both cases, we state we need to
rerun our scanning logic; otherwise, we can get stuck in a disconnected
state once the mode is switched back.