feat: update parser to fully handle lists
Previously, we updated our parser to handle the 'expansion' of variables
in lists. However, we had not updated the other two components: value
'evaluation' (where we import classes/functions and instantiate them if
of the form a.b.c() or a.b.c(some things)), and 'instantiation' (where
we instantiate classes/functions stored in dicts, with a 'class' key and
the rest of the keys corresponding to their kwargs).

I understand that these two are somewhat similar, but that's just how
the parser ended up... we can consider refactoring later, if necessary.

This CL updates list handling to support both 'evaluation' and 'instantiation'.