feat: visualizer takes in lists instead of maps
We were feeding maps to visualizer, which was all hunkey-dorey until we
realized that our parser cannot really 'expand' keys in a dict. The way
the parser works, we expand values of key:val pairs, since the keys are
generally presumed to correspond to input arguments (and thus should not
be expanded). On top of this: to instantiate classes/functions,
we instantiate everything when split into its own process. Thus, any
variable expansion will possibly turn a string into a larger dict! Since
dictionaries are not hashable (partially because hasable items should be
immutable, and dicts are explicitly *not* immutable).

Thus, inputting maps does not quite work. TO work around this, we simply
turn everything into lists! A list of keys, a list of origins, a list of
sizes, etc. In the constructor, we do a quick assert to ensure all lists
are of the same size, and then we're rolling!

This is likely not the cleanest solution, but it works and was
relatively fast to fix.