fix: fix set up viz for visualizer
In a previous CL, I pulled _set_up_visualization() into run_per_loop()
by mistake, instead of overriding run()! The idea was to set up the
visualization (i.e. create the figures) *after* the constructor, in the
run loop. But putting it in run_per_loop() was an obvious error. Fixing here.