feat: decouple visualizer from roi_experimenter
In the original roi_experimenter design, it was a child of Visualizer,
and started it up within it. This *does not* fit with our goal of clear,
concise components: we have a component doing 2 things!

So, to be more concise, we are splitting it up. We make a simple
reset_scans_visualizer, which resets its scans cache every N
frames (this is part of the logic roi_experimenter was doing). Then, we
clean up roi_experimenter to remove the visualization logic. Lastly, we
update config.toml in our sample to use this visualizer.