fix: update order of components in config example
Basically: the order they are written is the order they will be spawned.
Ideally, we want the device controller/afspm controller to be spawned
last, so we don't start running our experiment before our visualizer and
other guys are started up.

Note that this doesn't change *much*: if the experiment component is
started last, nothing will really happen until it is started up (since
it drives the experiment). However, it looks a bit weird to have work
being done *before* our visualizer is up and running. That is what
prompted me to change the order here.