fix: move trace logic to __init__.py
I had placed the 'trace' setup in spawn.py, but this means that we risk
failures if a submodule is loaded (since the only location where trace
was being set up was in spawn.py, which may not have been loaded).

By placing the trace 'setup' in the top-level __init__.py, we (I think)
guarantee that it will be set up for any submodule that is loaded.

For clarity, we also rename utils/logging.py to utils/log.py. Here, we
are trying to avoid confusing it with the main logger module, logger.py!