fix: cleaning up components logic tests
These were failing intermittently. After reviewing furhter, there were a
couple of clear issues:
- timeouts were not well designed. The afspmcon timeout should be longer
than that of the devcon (to account for time to get a reply); similarly,
the client timeout shoudl be longer than the afspmcon.
- updated startup_flush_messages() logic to poll a set number of times.
Previously, it flushed until it *first* received a message. However, if
more than one message was sent on startup, this would cause failures!