fix: set good timing defaults throughout project
In this CL, we do some relatively major refactoring in order to provide
sane defaults to all of the timing related variables (e.g.
poll_timeout_ms). It is quite easy to break this system if you
improperly set it up, so we figured we would allow the user to not worry
about it unless they explicitly want to.

To this end, we have added a set of defaults in io/common.py:
- REQUEST_TIMEOUT_MS: how long to wait between control requests. We
default this to 250 ms, which seems to be a sane default for TCP/IP
connections. Note that this may need to be increased if communicating
outside of the localhost...
- POLL_TIMEOUT_MS: how long to wait when poling, presumably within a
self-managed loop. This makes sense for the pubsub I/O path, since we
are often just polling to see if a message has been received.
- LOOPS_LEEP_S: how long to sleep between loops in our self-managed
loop. We set it to 100 ms, to allow some downtime for the PC to do
nothing.
- HEARTBEAT_PERIOD_S: how long to wait between beats of our
heartbeaters. We set it to 5s for now, although it could easily be much
longer.
- BEATS_BEFORE_DEAD: how many beat 'times' we will wait for before we
decide a component is dead. Set to 5 for now.

The biggest change here is that all of these timeout parameters now live
within their I/O components, and are removed from any components. Note
that this now means we explicitly differentiate between different types
of timeouts: request timeouts (rep-reqs) and poll timeouts (within
loops). This means we have more timeouts in intermediate I/O components,
like pubsubcache and control_router. However, this makes more sense: we
are setting proper defaults and giving the user more control over the
different timeouts.

Additionally, as they have defaults, they are often pulled to the end of
the input arguments of the associated constructors.

In addition to this large set of changes, we must mention some larger
fixes we found:
- We now have a startup sleep whenever we initialize a ZeroMQ component,
since we have found that a small startup is needed to avoid weird race
condition errors (us calling/sending messages before the socket is not
properly setup).
- For now, we have removed our general exception catching in
afspm_component / afspm_components_monitor, since it is not logging the
call stack. We will need to fix this later.