fix: move 'channel' from scanparams to scan2d
Previously, we had the name/channel parameter in ScanParameters2d.
Doing so means that we would be defining the 'channel' of interest for
scanning, which does not make much sense.

When running a scan, we will likely be running in a particular
'operating mode' (e.g. FM-AFM), which implies a number of channels being
recorded. More realistically, we would *subscribe* to a particular
channel (or set of channels) via Scan2d, rather than expliciting a
channel when setting scan parameters.

Because of this, in this CL we move the name parameter from the scan
params to the scan itself. Everything else in this CL is updating tests
for this change.