feat: add REQ_PARAM logic to control I/O
In this CL, we make some reasonably large changes, in order to allow the ability
to set/get certain parameters with the device controller.

... This is the point where the fancy design starts breaking. But hopefully
just a bit.

In order to do this, we:
- Create a params.py, holding a string enum with our device-agnostic setting
ids (as well as descriptions).
- Add a ParameterMsg type to control, the new REQ_PARAM request type, and a
couple of ControlResponse errors.
- Modify control/commands.py so that our responses can *include* objects (the
same as requests). To clean up/standardize, we also make the associated
methods match: parse_request()/parse_response() and serialize_request()/
serialize_response().
- Update ControlClient, ControlRouter, and ControlServer to account for this
change to replies. It gets a bit ugly, because now we have some methods where
the return value is either a response *or* a tuple (response, obj)...But such
is life, I suppose.
- Modify DeviceController to support set/get params, contain a param_method_map
(that maps parameters to methods to set/get them), and add a large description
trying to help clear up the *scope* of DeviceController.
- Add REQ_PARAM tests accordingly.