refactor: spurious kwargs, beat_period_s fix
1. There were many base classes with kwargs, when it didn't really make
sense/may be hiding bugs. It would make sense for a *child class* to use
kwargs, in order to not re-declare parameters that the base class
requires (although some may argue this is not good practice?). But it is
strange for it to happen with base classes! So, we fix a number of base
classes to do so.

Note that, for this to work, we also need to modify spawn.py's
_filter_requested_components() to purge the 'component' key from the
spawned class (as this is a spurious keyword argument that was not being
used, thanks to kwargs).

2. In a previous CL, we renamed hb_period_s to beat_period_s. However,
we forgot to make this change everywhere. Doing so here.