fix: adding AfspmController to gxsm testing
In the initial gxsm testing layout, we were just spawning the device
controller. However, for testing purposes it may make more sense to have
the afspm controller. Particularly, doing so will ensure there is a
cache, so that a test can be run whenever device controller + afspm
controller are detected in the 'testing' format; and it allows us to
introduce a 'skipif' to our test.

If we don't do this, we don't have an easy way to disable this test.
Pytest also does not seem to offer a way to manually enable a skipped
test, so not having this check may be difficult.

The alternative would be to define this 'test' as an experiment in and
of itself, with printed results...