feat: update spawn to not *require* monitor key
Previously, we required an 'afspm_components_monitor' key in our
config, from which we grabbed the appropriate constructor arguments.
However, all of these arguments are optional! Thus, it's a bit silly to
require this.

In this CL, we no longer require it. We pass the dict if it exists, but
otherwise do not.

We also update our main sample to no longer have this key.

Bonus: we remove explicit_config.toml and simplest_config.toml, as it
just confuses things (in my opinion).