fix: deep copy params when spawning stuff
It turns out that we were modifying the params_dict when spawning
components and listeners, because the dict is shared throughout. Thus,
if a component crashed and was restarted it could crash! Note that this
is recent, as we have removed the **kwargs that allow unnecessary
arguments to be passed to base class constructors.

To fix this, we are doing a deep copy of the params dict every time we
spawn a component or a listener.