fix: moving controller test to avoid auto-test
we move test_controller into samples/testing, to clearly distinguish it
from standard unit tests. Additionally, we add a pytest.inin, that tells
pytest to not recurse through samples/ when searching for tests.

This is all done to clearly distinguish our traditional unit tests with
our helper tests to validate a device controller works as expected.