fix/feat: make gxsm spawn monitorless (own script)
It appears we may need to spawn the gxsm DeviceController in a
monitorless fashion, due to peculiarities with how gxsm allows you to
run Python scripts (see spawn_gxsm.py description).

Here, we do that *and* pull the 'main'/spawning part into its own script
name (to presumably make it easier to distinguish).