fix: attempt to switch controller test for devcon
Our previous version of test_controller included an AfspmController in
the mix! We did this for one main reason: we were using pytest's
skipif() method, and would check if a DeviceController existed by
connecting a subscriber and seeing if we received data. Since the cache
logic to 'hold' memory is part of the AfspmController, we needed to have
one.

Now that we have made it so test_controller is not run by default when
testing, we don't really need that skipif() (and as such, removed it).
Thus, we change our tests to no longer use the afspmcon, since this is
kind of unnecessary.