feat: return to afspmcon in gxsm testing
I realize now that we kind of *need* an AfspmController in our tests, to
properly validate that 'initialization' is done properly. This is
because we need a 'cache' to be running, to allow a late connector (the
test) to receive the starting-state info.

It is also somewhat essential for test_scan_params, as we use the
initial scan params to define what we will try and set (to be more
likely we do not send parameters that are wrong).

The solution, for now, is to have an AfspmController running,
exclusively to manage the cache. This means that we have our test
connect its subscriber to the PubSubCache instance, but its client
connects directly to the DeviceController! I suppose this is not
necessary, and maybe confuses things. I may, in a follow-up CL, change
the tests to go through the AfspmController (meaning each test will
request control).