fix: better file handling with gxsm controller
GXSM may feed file paths that don't exist via chfname! Thus, we have to
validate that the fname provided exists on drive before adding it to our
list.

Also, we wrap gxsmread.read.open_dataset() in a try/catch. It is
crashing for reasons we must still investigate (likely competing NetCDF
versions), and we should probably account for crashes of gxsmread.