feat: make extract_proto support 'base' keys
Up until now, apparently, extract_proto() did not support 'base' keys if
you were using a more complicated cache logic. What I mean by this: the
'base' most key/envelope would be the message type name (i.e., what is
done in CacheLogic.get_envelope_for_proto).

In PBCCacheLogic, we intended to allow you to set envelopes in this way
as well! However, we forgot to account for it in extract_proto(). This
method was previously directly using the provided envelope to try and
access the associated proto. If you had *set up* your PBCCacheLogic with
'base' envelopes (e.g. Scan2d) but you *received* complicated
envelopes (e.g. Scan2d_Topo-Xp_36.0), it would simply crash with a
KeyError.

Now, we will (i) check to see if the envelope provided exists in
self.envelope_to_proto_Map; and (ii) if not, we try to see if a 'base'
one exists, by checking if any keys in envelope_to_proto_Map are
substrings of the provided envelope. If we find a substring match, we
use that 'base' envelope instead.