feat: prep for multi-subs and privatise I/O attrs
Two big things:
1. We add a ComboSubscriber, to allow you to bundle multiple
subscriptions into one equivalent Subscriber. For this, we modify
poll_and_store() to return either a *list* of (str, Message) or None.
Also, we make the attributes cache and shutdown_was_requested
properties (to alow usage with the ComboSubscriber).
2. For all of the I/O components, we become much more explicit about
internal/'private' attributes. Basically, since the whole point of these
I/O classes is to encapsulate all of the zmq logic, we make everything
except simple accessor methods internal.

(Note that we do not do this for the components level of classes. I am
still thinking about whether this is worth it.)

We add a test_scan_handler (for ScanHandler), and a test for ComboSubscriber.