feat: add AfspmComponentBase
We are renaming AfspmComponent as AfspmComponentBase, which centers on the
base functionality associated with a component. In the same package, we add
a *new* AfspmComponent, which is the base plus an optional publisher and
methods to be called on on_message_received() and run_per_loop(). Effectively,
this allows people to create AfspmComponents that function just based on
provided I/O nodes and methods (i.e. no new class definition necessary).

Additionally, we pulled get_heartbeat_url() from component.py into
heartbeat.py. We also make all internal classes inherit from AfspmComponentBase
instead of AfspmComponent.