feat: try to modify parsing approach [BROKEN]
The idea here was to modify how we parse the config file, to ensure that
referenced variables were instantiated *once* per component. This would
allow you to reference a variable in multiple places, and have their
state be stored for a component.

I couldn't get it working. It ended up being fairly complicated,
where we split up components and variables, expand and instantiate the
variables per component, and then try and expand them into the
component (which we finally instantiate).

The issue appears to be linked to our merging of the two dicts. At some
point, the zmq sockets are cloned; when we call them outside of the
construct method, they fail...