fix: Get consistent proto IDs for scans
Previously, we were directly feeding the ROI size to our method
get_envelope_for_proto() in PBCScanLogic. However, depending on our
float math, we might get a value that is not *exactly* the same (e.g.
39.9999... instead of 40).

The solution (I think), is to ensure we do a rounding operation on our value.