feat: auto-set ControlClient UUID and merge kwargs
Two things:
1. For a while now, we have been forced to *explicit* the ControlClient
UUIDs in our configs. To avoid this hassle, we now have AfspmComponent
automatically set the UUID to its name attribute, if
override_client_uuid is true (and a ControlClient was provided). To do
this, we add a method set_uuid() in ControlClient.
2. In order to allow a common object to be fed to the
message_received_method and per_loop_method, we use one singular kwargs
now.

Why must we do (2)? Well, when we are parsing our config file, we
expand all our variables *at the beginning*, and only instantiate them
per-component before starting up our component. Thus, if a variable is
referenced more than 1x by a component, we will be spawning a *new*
instance for each of those references!

This is an annoying issue. I spent a while developing a larger solution,
where I modified how our parser works, and expand later on in the
parsing process. However, I hit a different roadblock: my instantiated
zmq sockets were not behaving properly. I have this 'solution' stashed
locally, so I may return to it if I get the time.