feat: add SS_BUSY_PARAM state
Here, we add a busy param state, which can be optionally used to
indicate that a REQ_PARAM request that sets a parameter will take a bit
to get ready to scan. We do no special checks/validation/updates out of
this scan state, so you *must* make sure your poll_scan_state() method
handles this logic for any method you are considering.