feat: add ZCtrlParameters and pubsub/control
At a high-level, we are adding a ZCtrlParameters structure, to allow
subscribers to see Z-control feedback changes, and contorl clients to
easily set them. Note that we only control feedbackOn, and P- and I-
gains, (no setpoint or error gains). We do this because we believe only
these settings are necessary for any feedback automation.

- We add a ZCtrlParameters message type, in feedback.proto, and
- REQ_SET_ZCTRL_PARAMS to set it.
- We update device/controller.py and control/client.py to support
- setting ZCtrlParameters.
- We update device/params.py, to remove these attributes.
- We update sample_components.py (testing) and image/image_controller.py
- to implement ZCtrl abstract methods.
- We update cache_logic.py's default proto-with-hist seq to add zctrlparameters.